refactor DDR4Interface
