+++ src/DRAMPower/DRAMPower/standards/ddr4/DDR4Interface.cpp
@@ -2,6 +2,44 @@
 
 namespace DRAMPower {
 
+    DDR4Interface::DDR4Interface(const MemSpecDDR4 &memSpec, implicitCommandInserter_t&& implicitCommandInserter, 
+                  patternHandler_t &patternHandler)
+    : m_commandBus{cmdBusWidth, 1,
+        util::BusIdlePatternSpec::H, util::BusInitPatternSpec::H}
+    , m_dataBus{
+        util::databus_presets::getDataBusPreset(
+            memSpec.bitWidth * memSpec.numberOfDevices,
+            util::DataBusConfig {
+                memSpec.bitWidth * memSpec.numberOfDevices,
+                memSpec.dataRate,
+                util::BusIdlePatternSpec::H,
+                util::BusInitPatternSpec::H,
+                DRAMUtils::Config::TogglingRateIdlePattern::H,
+                0.0,
+                0.0,
+                util::DataBusMode::Bus
+            }
+        )
+    }
+    , m_readDQS(memSpec.dataRate, true)
+    , m_writeDQS(memSpec.dataRate, true)
+    , m_clock(2, false)
+    , m_dbi(memSpec.numberOfDevices * memSpec.bitWidth, util::DBI::IdlePattern_t::H, 8,
+        [this](timestamp_t load_timestamp, timestamp_t chunk_timestamp, std::size_t pin, bool inversion_state, bool read) {
+        this->handleDBIPinChange(load_timestamp, chunk_timestamp, pin, inversion_state, read);
+    }, false)
+    , m_dbiread(m_dbi.getChunksPerWidth(), util::Pin{m_dbi.getIdlePattern()})
+    , m_dbiwrite(m_dbi.getChunksPerWidth(), util::Pin{m_dbi.getIdlePattern()})
+    , prepostambleReadMinTccd(memSpec.prePostamble.readMinTccd)
+    , prepostambleWriteMinTccd(memSpec.prePostamble.writeMinTccd)
+    , m_ranks(memSpec.numberOfRanks)
+    , m_memSpec(memSpec)
+    , m_patternHandler(patternHandler)
+    , m_implicitCommandInserter(std::move(implicitCommandInserter))
+    {
+        registerPatterns();
+    }
+
     void DDR4Interface::registerPatterns() {
         using namespace pattern_descriptor;
         m_patternHandler.get().registerPattern<CmdType::ACT>({
