+++ src/DRAMPower/DRAMPower/util/RegisterHelper.h
@@ -67,6 +67,33 @@ private:
     std::vector<Rank> &m_ranks;
 };
 
+template<typename Interface>
+struct InterfaceRegisterHelper {
+// Public constructors and assignment operators
+public:
+    InterfaceRegisterHelper(Interface *interface)
+        : m_interface(interface)
+    {}
+    InterfaceRegisterHelper(const InterfaceRegisterHelper&) = default; // copy constructor
+    InterfaceRegisterHelper& operator=(const InterfaceRegisterHelper&) = default; // copy assignment operator
+    InterfaceRegisterHelper(InterfaceRegisterHelper&&) = default; // Move constructor
+    InterfaceRegisterHelper& operator=(InterfaceRegisterHelper&&) = default; // Move assignment operator
+    ~InterfaceRegisterHelper() = default; // Destructor
+
+// Public member functions
+    template<typename Func>
+    decltype(auto) registerHandler(Func &&member_func) {
+        Interface *this_ptr = m_interface;
+        return [this_ptr, member_func](const Command & command) {
+            (this_ptr->*member_func)(command);
+        };
+    }
+
+// Private member variables
+private:
+    Interface *m_interface;
+};
+
 } // namespace DRAMPower::util
 
 #endif /* DRAMPOWER_STANDARDS_REGISTERHELPER_H */
